Add Bird token and use it for existing cards.
