+++ Mage.Sets/src/mage/sets/returntoravnica/SellerOfSongbirds.java
@@ -28,14 +28,14 @@
 package mage.sets.returntoravnica;
 
 import java.util.UUID;
-import mage.constants.CardType;
-import mage.constants.Rarity;
 import mage.MageInt;
 import mage.abilities.common.EntersBattlefieldTriggeredAbility;
 import mage.abilities.effects.common.CreateTokenEffect;
 import mage.abilities.keyword.FlyingAbility;
 import mage.cards.CardImpl;
-import mage.game.permanent.token.Token;
+import mage.constants.CardType;
+import mage.constants.Rarity;
+import mage.game.permanent.token.BirdToken;
 
 /**
  * @author LevelX2
@@ -62,18 +62,5 @@ public class SellerOfSongbirds extends CardImpl {
     @Override
     public SellerOfSongbirds copy() {
         return new SellerOfSongbirds(this);
-    }
-
-    private class BirdToken extends Token {
-
-        public BirdToken() {
-            super("Bird", "1/1 white Bird creature token with flying");
-            cardType.add(CardType.CREATURE);
-            color.setWhite(true);
-            subtype.add("Bird");
-            power = new MageInt(1);
-            toughness = new MageInt(1);
-            addAbility(FlyingAbility.getInstance());
-        }
     }
 }
